+++ renderer/src/widgets/font.rs
@@ -127,7 +127,7 @@ fn draw_ttf_text(
     let width = glyphs
         .iter()
         .rev()
-        .map(|g| g.position().x as f32 + g.unpositioned().h_metrics().advance_width)
+        .map(|g| g.position().x + g.unpositioned().h_metrics().advance_width)
         .next()
         .unwrap_or(0.0)
         .ceil() as usize;
